requirements.txt
# Core dependencies
PyQt5==5.15.9
pynput==1.7.6
numpy==1.24.3
sounddevice==0.4.6
pyaudio==0.2.13
azure-cognitiveservices-speech==1.30.0
google-cloud-speech==2.21.0
pyautogui==0.9.54
pyperclip==1.8.2
requests==2.31.0
python-dotenv==1.0.0
PyInstaller==5.13.0
pywin32>=300; platform_system=="Windows"
keyboard>=0.13.5
wave>=0.0.2
psutil>=5.9.0

# Speech recognition services
openai>=1.3.0
azure-cognitiveservices-speech==1.30.0
google-cloud-speech==2.21.0

# Local services
whisper>=20240930
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0

# Packaging and distribution
pyinstaller==6.0.0 